[JKM] qrcode 데이터 생성 메서드 위치를 좀더 알맞은 서비스 객체로 옮기자
