+++ app/src/main/java/com/emart24/service/QRService.kt
@@ -1,5 +1,6 @@
 package com.emart24.service
 
+import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Intent
 import android.graphics.Bitmap
@@ -8,8 +9,21 @@ import com.google.zxing.integration.android.IntentIntegrator
 import com.google.zxing.integration.android.IntentResult
 import com.google.zxing.qrcode.QRCodeWriter
 import com.journeyapps.barcodescanner.BarcodeEncoder
+import java.security.MessageDigest
+import java.util.*
 
 class QRService {
+
+    @SuppressLint("SimpleDateFormat")
+    fun generateQRCode(last4DigitsOfPhoneNumber: String, dateTime: String): String {
+        val sha256 = MessageDigest.getInstance("SHA-256")
+        val qrCodeByteArray = "$last4DigitsOfPhoneNumber$dateTime".toByteArray()
+        sha256.update(qrCodeByteArray)
+        val qrCode = Arrays.toString(sha256.digest())
+        println("QR Code: $qrCode, QR Code Size: ${qrCode.length}")
+        return qrCode
+    }
+
     fun createQR(qrcodeData: String?): Bitmap? {
         val writer = QRCodeWriter()
         val matrix = writer.encode(qrcodeData, BarcodeFormat.QR_CODE, 300, 300)
